Set initial values for auto-generated IDs to null
The initial values for the auto-generated primary keys in the entities `Saving`, `ExpenseCategory`, `Income`, `Budget`, `Expense`, and `SavingGoal` are modified from `0` to `null`. This change is made to correctly represent that these IDs are auto-generated and do not have a meaningful default value. The Android Room library handles auto-incrementing these fields.